pyproject.toml
[project]
name = "explore-mujoco"
version = "0.1.0"
description = "RL experiments with SO-101 arm in MuJoCo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "torchaudio>=2.8.0",
    "pytorch-triton-rocm>=3.2.0",
    "mujoco>=3.0.0",
    "gymnasium @ git+https://github.com/stepjam/Gymnasium.git@0.29.2",
    "stable-baselines3>=2.0.0",
    "imageio[ffmpeg]>=2.30.0",
    "tensorboard>=2.14.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.3",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",
    "numpy<2.0.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-rocm" },
]
torchvision = [
    { index = "pytorch-rocm" },
]
torchaudio = [
    { index = "pytorch-rocm" },
]
pytorch-triton-rocm = [
    { index = "pytorch-rocm" },
]

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true
